refactor(login): refatorando metodo de login e deletando imports nao usados
