+++ teste/src/app/app-routing.ts
@@ -2,7 +2,7 @@ import { MenuComponent } from '../app/menu/menu.component';
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router'; 
 import { LoginComponent } from './login/login.component';
-import { AuthGuard } from './login/auth.guard'
+import { AuthGuard } from './login/auth/auth.guard'
 import { Error404pageComponent } from './error404page/error404page.component'
 
 const APP_ROUTES: Routes = [
