+++ teste/src/app/login/login.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Usuario } from './usuario'
-import { AuthService } from './auth.service'
 import { Router } from '@angular/router';
+import { AuthService } from './auth/auth.service';
 
 @Component({
   selector: 'app-login',
@@ -12,20 +12,16 @@ export class LoginComponent implements OnInit {
 
   public usuario: Usuario = new Usuario()
 
-   constructor(private router: Router) { } 
+   constructor(private router: Router, private authService: AuthService) { } 
 
     ngOnInit() {    
       }
-
-    fazerLogin(){
-      //this.authService.fazerLogin(this.usuario)
-      if(localStorage['token'] = '123@qwe'){    
-    this.router.navigate(['home']);
-      }else{
-        this.router.navigate(['home']);
-      }
+      
+    fazerLogin(){ 
+     this.authService.fazerLogin(this.usuario)
+     
     }
-  }
+}
